GUI/requirement.txt
google-api-python-client
httplib2
numpy
oauth2client
Pillow
protobuf
pyserial
requests
opencv-python